+++ lib/flutter_value_localization.dart
@@ -1,3 +1,4 @@
 export 'value_localization.dart';
 export 'localization_widget_builder.dart';
 export 'system_lang_mapper.dart';
+export 'map_localization_source.dart';